Graphics.Heatmap

Description

Tools for generating heatmap images.

Documentation

type Point = V2 Natural

A Point as used throughout this heatmap codebase is currently limited to two dimensional vectors of natural numbers.

mkDot :: Point -> Natural -> [(Point, Depth)]

Make a dot of a given size indicating presence that should begin heating the map.

genHeatmap :: (Natural, Natural) -> Colorizer -> [(Point, Depth)] -> Image PixelRGBA8

Generate a heatmap image using a Colorizer and a collection of points and depths.

mkTranslator :: RealFrac a => (V2 a, V2 a) -> (V2 Natural, V2 Natural) -> V2 a -> V2 Natural

Make a scale translator that can translate points in an arbitrarily bounded two dimensional vector to our 2D vector with Natural bounds we use when applying points.