Graphics.Heatmap.Types

Description

Heatmap types.

Documentation

type Colorizer = Depth -> PixelRGBA8

A Colorizer produces a pixel for a given depth.

type ColorScheme = [PixelRGBA8]

A ColorScheme is a collection of pixels that Depths can map to.

newtype Depth

Depth represents how warm a particular spot is.

Depths as a semigroup may be added together, but can never have a value larger than maxBound.

Constructors

Depth Word8

alphaColorizer :: PixelRGB8 -> Colorizer

Simple colorizer that uses a constant color and applies different alpha values.

schemeColorizer :: ColorScheme -> Colorizer

Colorizer that applies a ColorScheme along with the depth.